+++ page-templates/exhibit.php
@@ -1,0 +1,14 @@
+<?php
+
+/**
+ * Template Name: Exhibit
+ *
+ * @package CIV Shows
+ */
+
+get_header();
+
+get_template_part('template-parts/pages/page-header-exhibit');
+get_template_part('template-parts/components/exhibit-with-us');
+
+get_footer();